@@ -14,29 +14,31 @@
     "build:dev:client": "node_modules/webpack/bin/webpack.js -w --config webpack.config.js"
   },
   "dependencies": {
-    "aphrodite": "^1.0.0",
+    "aphrodite": "^1.1.0",
     "autoprefixer": "^6.3.7",
-    "babel": "^6.5.2",
-    "babel-core": "^6.16.0",
-    "babel-loader": "^6.2.5",
+    "babel": "^6.23.0",
+    "babel-core": "^6.23.1",
+    "babel-loader": "^6.3.2",
     "babel-plugin-react-transform": "^2.0.2",
     "babel-plugin-transform-class-properties": "^6.18.0",
     "babel-plugin-transform-object-rest-spread": "^6.8.0",
     "babel-polyfill": "^6.9.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
+    "babel-preset-env": "^1.1.8",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-latest": "^6.22.0",
+    "babel-preset-react": "^6.23.0",
     "babel-preset-stage-0": "^6.5.0",
     "babel-relay-plugin": "^0.10.0",
     "babel-relay-plugin-loader": "^0.10.0",
     "classnames": "^2.2.5",
-    "css-loader": "^0.23.1",
-    "dialog-polyfill": "^0.4.5",
+    "css-loader": "^0.26.1",
+    "dialog-polyfill": "^0.4.7",
     "dompurify": "^0.8.4",
     "expose-loader": "^0.7.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "formsy": "^0.19.0",
-    "formsy-material-ui": "^0.5.3",
-    "formsy-react": "^0.19.0",
+    "extract-text-webpack-plugin": "^2.0.0-rc.3",
+    "formsy": "^0.19.2",
+    "formsy-material-ui": "^0.5.5",
+    "formsy-react": "^0.19.2",
     "highlight.js": "^9.9.0",
     "immutability-helper": "^2.0.0",
     "imports-loader": "^0.6.5",
@@ -44,22 +46,22 @@
     "jquery-ujs": "^1.2.2",
     "json-loader": "^0.5.4",
     "marked": "^0.3.6",
-    "material-ui": "^0.16.5",
+    "material-ui": "^0.17.0",
     "moment": "^2.17.1",
-    "node-sass": "^4.0.0",
+    "node-sass": "^4.5.0",
     "promise-window": "^1.1.0",
     "query-string": "^4.2.3",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
     "react-loader": "^2.4.0",
     "react-relay": "^0.10.0",
     "react-rte": "^0.11.0",
     "react-tap-event-plugin": "^2.0.1",
     "react-truncate": "^2.0.3",
-    "sass-loader": "^4.0.2",
+    "sass-loader": "^6.0.1",
     "style-loader": "^0.13.1",
     "underscore": "^1.8.3",
-    "webpack": "^1.14.0"
+    "webpack": "^2.2.1"
   },
   "license": "MIT",
   "engines": {
