@@ -8,7 +8,7 @@ import { sanitizeText } from '../../../utils/sanitize.es6';
 import '../../styles/pins.sass';
 
 const Data = (props) => {
-  const { item, toggleItemOnServer } = props;
+  const { item, unpinImport, pinImport } = props;
   const description = sanitizeText(
     item.description || item.body || item.summary || item.tagline
   );
@@ -18,6 +18,7 @@ const Data = (props) => {
       pr: 'comments',
       repo: 'stargazers_count',
     },
+    medium: 'recommends',
     stackoverflow: 'up_vote_count',
     youtube: 'likeCount',
     meetup: 'yes_rsvp_count',
@@ -35,7 +36,9 @@ const Data = (props) => {
         <Checkbox
           checked={item.pinned}
           style={{ top: 'calc(100% / 3)' }}
-          onCheck={event => toggleItemOnServer(event, item)}
+          onCheck={(event, checked) =>
+            (checked ? pinImport(event, item) : unpinImport(event, item))
+          }
         />
       }
       rightIcon={
@@ -78,7 +81,8 @@ const Data = (props) => {
 
 Data.propTypes = {
   item: React.PropTypes.object,
-  toggleItemOnServer: React.PropTypes.func,
+  pinImport: React.PropTypes.func,
+  unpinImport: React.PropTypes.func,
 };
 
 const DataContainer = Relay.createContainer(Data, {
@@ -94,6 +98,7 @@ const DataContainer = Relay.createContainer(Data, {
         stargazers_count,
         likeCount,
         viewCount,
+        recommends,
         yes_rsvp_count,
         summary,
         up_vote_count,
