@@ -5,22 +5,38 @@ import React, { Component } from 'react';
 import Relay from 'react-relay';
 import { Card, CardMedia, CardActions, CardTitle, CardText } from 'material-ui/Card';
 import moment from 'moment';
-import IconButton from 'material-ui/IconButton';
 import FontIcon from 'material-ui/FontIcon';
 
 // Child Components icons
 import YoutubeIcon from '../../shared/icons/youtube.es6';
 import { sanitizeText } from '../../../utils/sanitize.es6';
+import AchievementForm from './form.es6';
+import AchievementActions from './actions.es6';
 
 class Youtube extends Component {
+  constructor(props) {
+    super(props);
+    this.edit = this.edit.bind(this);
+    this.state = {
+      editing: false,
+    };
+  }
+
   componentDidMount() {
     setTimeout(() => {
       this.iframe.setAttribute('src', this.iframe.getAttribute('data-src'));
     }, 1000);
   }
 
+  edit(event) {
+    if (event) {
+      event.preventDefault();
+    }
+    this.setState({ editing: !this.state.editing });
+  }
+
   render() {
-    const { achievement, remove } = this.props;
+    const { achievement, remove, update } = this.props;
     const embedVideoStyle = {
       display: 'block',
       width: '100%',
@@ -42,57 +58,63 @@ class Youtube extends Component {
                 </h2>
 
                 {achievement.is_owner ?
-                  <IconButton
-                    className="remove"
-                    tooltip="Remove"
-                    tooltipStyles={{ top: 25 }}
-                    onClick={event => remove(event, achievement)}
-                  >
-                    <FontIcon className="material-icons">close</FontIcon>
-                  </IconButton> : ''
+                  <AchievementActions
+                    achievement={achievement}
+                    remove={remove}
+                    edit={this.edit}
+                  /> : ''
                 }
 
                 <time className="date">
                   {
-                    moment(achievement.created_at, 'YYYY-MM-DD HH:mm:ss [UTC]')
+                    moment(achievement.date, 'YYYY-MM-DD HH:mm:ss [UTC]')
                     .format('MMMM Do YYYY')
                     .toString()
                   }
                 </time>
 
-                <CardMedia
-                  className="achievement-card-media"
-                >
-                  <div className="video-embed" style={embedVideoStyle}>
-                    <iframe
-                      ref={node => (this.iframe = node)}
-                      style={embedVideoStyle}
-                      frameBorder="0"
-                      data-src={`//youtube.com/embed/${achievement.source_id}`}
+                {this.state.editing ?
+                  <AchievementForm
+                    achievement={achievement}
+                    update={update}
+                    edit={this.edit}
+                  /> :
+                  <div className="achievement-content">
+                    <CardMedia
+                      className="achievement-card-media"
+                    >
+                      <div className="video-embed" style={embedVideoStyle}>
+                        <iframe
+                          ref={node => (this.iframe = node)}
+                          style={embedVideoStyle}
+                          frameBorder="0"
+                          data-src={`//youtube.com/embed/${achievement.source_id}`}
+                        />
+                      </div>
+                    </CardMedia>
+
+                    <CardTitle
+                      className="achievement-card-header"
+                      title={
+                        <div className="title">
+                          {achievement.title}
+                        </div>
+                      }
+                    />
+                    <CardText
+                      className="achievement-card-description"
+                      dangerouslySetInnerHTML={{
+                        __html: sanitizeText(achievement.description),
+                      }}
                     />
                   </div>
-                </CardMedia>
-
-                <CardTitle
-                  className="achievement-card-header"
-                  title={
-                    <div className="title">
-                      {achievement.title}
-                    </div>
-                  }
-                />
-                <CardText
-                  className="achievement-card-description"
-                  dangerouslySetInnerHTML={{
-                    __html: sanitizeText(achievement.description),
-                  }}
-                />
+                }
 
                 <CardActions className="meta">
                   <span className="badge">
                     {`${achievement.likeCount}`}
                     <FontIcon
-                      color="#fff"
+                      color="#333"
                       className="material-icons"
                       style={{
                         fontSize: 20,
@@ -107,7 +129,7 @@ class Youtube extends Component {
                   <span className="badge">
                     {`${achievement.viewCount}`}
                     <FontIcon
-                      color="#fff"
+                      color="#333"
                       className="material-icons"
                       style={{
                         fontSize: 20,
@@ -131,24 +153,26 @@ class Youtube extends Component {
 Youtube.propTypes = {
   achievement: React.PropTypes.object,
   remove: React.PropTypes.func,
+  update: React.PropTypes.func,
 };
 
 const YoutubeContainer = Relay.createContainer(Youtube, {
   fragments: {
     achievement: () => Relay.QL`
-      fragment on Import {
+      fragment on Achievement {
         id,
         title,
         source_name,
         source_id,
         description,
         developer_id,
-        connection_id,
+        import_id,
         is_owner,
         likeCount,
         viewCount,
-        pinned,
-        created_at,
+        date,
+        ${AchievementActions.getFragment('achievement')},
+        ${AchievementForm.getFragment('achievement')},
       }
     `,
   },
