@@ -16,9 +16,12 @@ import PopupComposer from '../email/popupComposer.es6';
 import composerRoute from '../../routes/composerRoute.es6';
 import LoadingComponent from '../shared/loadingComponent';
 import ErrorComponent from '../shared/errorComponent';
+import CurrentUser from '../../helpers/currentUser.es6';
 
 // Stylesheet
 import iconStyles from '../styles/icons.es6';
+
+const currentUser = new CurrentUser();
 
 class Links extends Component {
   static openUrl(event, url) {
@@ -74,10 +77,10 @@ class Links extends Component {
             this.props.developer.email &&
             this.props.developer.hireable ?
               <Chip
-                labelStyle={{ fontSize: 14 }}
+                labelStyle={{ fontSize: 14, fontWeight: 500 }}
                 className={css(iconStyles.linkIcon, iconStyles.hover)}
                 onClick={event =>
-                  (this.props.developer.premium ?
+                  (currentUser.isEmployer && this.props.developer.premium ?
                     this.openComposer(event) : this.openMail(event))
                 }
                 style={{ cursor: 'pointer' }}
@@ -97,7 +100,7 @@ class Links extends Component {
 
           {this.props.developer.blog ?
             <Chip
-              labelStyle={{ fontSize: 14 }}
+              labelStyle={{ fontSize: 14, fontWeight: 500 }}
               className={css(iconStyles.linkIcon, iconStyles.hover)}
               style={{ cursor: 'pointer' }}
               onClick={event => Links.openUrl(event, this.props.developer.blog)}
@@ -115,7 +118,7 @@ class Links extends Component {
           }
 
           <Chip
-            labelStyle={{ fontSize: 14 }}
+            labelStyle={{ fontSize: 14, fontWeight: 500 }}
             className={css(iconStyles.linkIcon, iconStyles.hover)}
             style={{ cursor: 'pointer' }}
             onClick={event => Links.openUrl(event, this.props.developer.html_url)}
@@ -133,7 +136,7 @@ class Links extends Component {
 
           {this.props.developer.linkedin ?
             <Chip
-              labelStyle={{ fontSize: 15, paddingLeft: 5 }}
+              labelStyle={{ fontSize: 14, fontWeight: 500, paddingLeft: 5 }}
               className={css(iconStyles.linkIcon, iconStyles.hover)}
               style={{ cursor: 'pointer' }}
               onClick={event => Links.openUrl(event, this.props.developer.linkedin)}
