@@ -3,7 +3,7 @@ import createDOMPurify from 'dompurify';
 export function sanitizeRichText(text) {
   return createDOMPurify.sanitize(text, {
     ALLOWED_TAGS: [
-      'b', 'i', 'em', 'ins', 'code', 'p', 'li', 'ul', 'ol', 'strong',
+      'b', 'i', 'em', 'ins', 'code', 'p', 'li', 'ul', 'ol', 'strong', 'a',
     ],
   });
 }
