@@ -1,6 +1,6 @@
 module Developers
-  ToggleAchievementMutation = GraphQL::Relay::Mutation.define do
-    name 'ToggleAchievement'
+  PinAchievementMutation = GraphQL::Relay::Mutation.define do
+    name 'PinAchievement'
     description 'Add or remove an achievement from profile based on selection'
 
     # Define input and return field
@@ -8,9 +8,10 @@ module Developers
 
     # Return field
     return_field :developer, DeveloperType
-    return_field :connection, ConnectionType
+    return_field :import, ImportType
+    return_field :achievementEdge, AchievementType.edge_type
 
     # Resolve block to update a model
-    resolve(Developers::ToggleAchievementResolver)
+    resolve(Developers::AddAchievementResolver)
   end
 end
