@@ -1,28 +1,23 @@
-ImportType = GraphQL::ObjectType.define do
-  name 'Import'
-  description 'Fetch import associated fields'
+AchievementType = GraphQL::ObjectType.define do
+  name 'Achievement'
+  description 'Fetch achievement associated fields'
   interfaces [GraphQL::Relay::Node.interface]
   global_id_field :id
 
   field :source_id, types.String, 'Data source id'
   field :category, types.String, 'Data source category'
   field :developer_id, types.String, 'Data owner id'
-  field :connection_id, types.String, 'Data source connection id'
 
-  field :database_id, types.Int, 'Import database id' do
-    resolve ->(obj, _args, _ctx) do
-      obj.id
+  field :import_id, types.ID, 'Data import id' do
+    resolve ->(obj, _args, ctx) do
+      Schema.id_from_object(obj.import, ImportType, ctx)
     end
   end
 
   field :source_name, types.String, 'Data source name'
   field :title, types.String, 'title of this data source'
-  field :name, types.String, 'Name of this data source'
   field :description, types.String, 'Description of this data source'
-  field :summary, types.String, 'Position summary(in)'
-  field :tagline, types.String, 'Product tagline(ph)'
   field :featured, types.Boolean, 'Product featured(ph)'
-  field :body, types.String, 'Body of this data source'
   field :stargazers_count, types.Int, 'Total stars for this data source'
   field :likeCount, types.Int, 'Total likes for this data source'
   field :up_vote_count, types.Int, 'Total up votes for this data source'
@@ -30,16 +25,15 @@ ImportType = GraphQL::ObjectType.define do
   field :comments_count, types.Int, 'Total comments (ph)'
   field :votes_count, types.Int, 'Total votes count (ph)'
   field :recommends, types.Int, 'Total recommends count (medium)'
+  field :mediumUrl, types.String, 'Medium URL'
   field :yes_rsvp_count, types.Int, 'Total attendess for event(meetup)'
   field :viewCount, types.Int, 'Total views (youtube)'
   field :comments, types.Int, 'Total comments (github pr)'
-  field :html_url, types.String, 'Url of this source(github)'
   field :is_accepted, types.Boolean, 'Is answer accepted?(stackoverflow)'
   field :isCurrent, types.Boolean, 'Is current job? (in)'
   field :language, types.String, 'Language of this repo'
-  field :link, types.String, 'Link of this source(so)'
-  field :created_at, types.String, 'When this item was created'
-  field :pinned, types.Boolean, 'Is answer pinned?'
+  field :link, types.String, 'The link of this item'
+  field :date, types.String, 'When this item was created'
   field :discussion_url, types.String, 'Discussion url (ph)'
   field :redirect_url, types.String, 'Product url (ph)'
 
