@@ -1,5 +1,5 @@
 # This migration comes from mailboxer_engine (originally 20110511145103)
-class CreateMailboxer < ActiveRecord::Migration
+class CreateMailboxer < ActiveRecord::Migration[4.2]
   def self.up
   #Tables
     #Conversations
